src/main/java/RequestTrackerServletFilter.java
import org.slf4j.MDC;

import javax.servlet.Filter;
import javax.servlet.FilterChain;
import javax.servlet.FilterConfig;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.UUID;

public class RequestTrackerServletFilter implements Filter {
	@Override
	public void init(FilterConfig filterConfig) throws ServletException {

	}

	@Override
	public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
		String requestId = ((HttpServletRequest)request).getHeader("X-Request-Tracker");

		MDC.put("Request-Tracker", UUID.randomUUID().toString());

		chain.doFilter(request, response);
	}

	@Override
	public void destroy() {

	}
}
